Manual page!
For now it depends on [ronn](https://github.com/rtomayko/ronn) being
installed. If `ronn` is not available, `make manual` and `make install`
will fail.

Eventually we should be able to not rely on software that requires the
entire ruby toolchain (and then some!) to be able to build mpp. But for
now, it will do.